requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
tensorflow
keras
jupyter
Pillow
streamlit

